backend/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
